Numeric.EMD

Description

Empirical Mode Decomposition in pure Haskell.

Main interface is emd, with defaultEO. A tracing version that outputs a log to stdout is also available, as emdTrace. This can be used to help track down a specific IMF that might be taking more time than desired.

This package uses "sized vectors" as its main interface, to ensure:

The resulting EMD contains IMFs that are all the same length as the input vector

We provide a vector of size of at least one.

There are many functions to convert unsized vectors to sized vectors in Data.Vector.Sized and associated modules, including toSized (for when you know the size at compile-time) and withSized (for when you don't).

Empirical Mode Decomposition

emd :: (Vector v a, KnownNat n, Floating a, Ord a) => EMDOpts v (n + 1) a -> Vector v (n + 1) a -> EMD v (n + 1) a

EMD decomposition of a given time series with a given sifting stop condition.

Takes a sized vector to ensure that:

The resulting EMD contains IMFs that are all the same length as the input vector

We provide a vector of size of at least one.

emdTrace :: (Vector v a, KnownNat n, Floating a, Ord a, MonadIO m) => EMDOpts v (n + 1) a -> Vector v (n + 1) a -> m (EMD v (n + 1) a)

emd, but tracing results to stdout as IMFs are found. Useful for debugging to see how long each step is taking.

emd' :: (Vector v a, KnownNat n, Floating a, Ord a, Applicative m) => (SiftResult v (n + 1) a -> m r) -> EMDOpts v (n + 1) a -> Vector v (n + 1) a -> m (EMD v (n + 1) a)

emd with a callback for each found IMF.

iemd :: (Vector v a, Num a) => EMD v n a -> Vector v n a

Collapse an EMD back into its original time series. Should be a left-inverse to emd: using iemd on the result of emd should give back the original vector.

Since: 0.1.5.0

data EMD v n a

An EMD v n a is an Empirical Mode Decomposition of a time series with n items of type a stored in a vector v.

The component-wise sum of emdIMFs and emdResidual should yield exactly the original series (see iemd).

Constructors

EMD

Fields

emdIMFs :: ![Vector v n a]

emdResidual :: !(Vector v n a)

Configuration

data EMDOpts v n a

Options for EMD composition.

Constructors

EO

Fields

eoSifter :: Sifter v n a

stop condition for sifting

eoSplineEnd :: SplineEnd a

end conditions for envelope splines

eoBoundaryHandler :: Maybe BoundaryHandler

process for handling boundary

defaultEO :: (Vector v a, Fractional a, Ord a) => EMDOpts v n a

Default EMDOpts

Note: If you immediately use this and set eoSifter, then v will be ambiguous. Explicitly set v with type applications to appease GHC

defaultEO @(Data.Vector.Vector)
   { eoSifter = scTimes 100
   }

data BoundaryHandler

Boundary conditions for splines.

Constructors

BHClamp
Clamp envelope at end points (Matlab implementation)

BHSymmetric
Extend boundaries symmetrically

data Sifter v n a

A sift stopping condition.

It is a Pipe consumer that takes single sift step results upstream and terminates with () as soon as it is satisfied with the latest sift step.

Use combinators like siftOr and siftAnd to combine sifters, and the various sifters in Numeric.EMD.Sift to create sifters from commonly established ones or new ones from scratch.

Since: 0.2.0.0

defaultSifter :: (Vector v a, Fractional a, Ord a) => Sifter v n a

Default Sifter

defaultSifter = siftStdDev 0.3 siftOr siftTimes 50

R package uses siftTimes 20, Matlab uses no limit

data SplineEnd a

End condition for spline

Constructors

SENotAKnot
"Not-a-knot" condition: third derivatives are continuous at endpoints. Default for matlab spline.

SENatural
"Natural" condition: curve becomes a straight line at endpoints.

SEClamped a a
"Clamped" condition: Slope of curves at endpoints are explicitly given.

Since: 0.1.2.0

Internal

sift :: forall v n a. (Vector v a, KnownNat n, Floating a, Ord a) => EMDOpts v (n + 1) a -> Vector v (n + 1) a -> SiftResult v (n + 1) a

Iterated sifting process, used to produce either an IMF or a residual.

data SiftResult v n a

The result of a sifting operation. Each sift either yields a residual, or a new IMF.

Constructors

SRResidual !(Vector v n a)

SRIMF !(Vector v n a) !Int
number of sifting iterations

envelopes :: (Vector v a, KnownNat n, Fractional a, Ord a) => SplineEnd a -> Maybe BoundaryHandler -> Vector v (n + 1) a -> Maybe (Vector v (n + 1) a, Vector v (n + 1) a)

Returns cubic splines of local minimums and maximums. Returns Nothing if there are not enough local minimum or maximums to create the splines.

Orphan instances

(Vector v a, Fractional a, Ord a) => Default (EMDOpts v n a)
Since: 0.1.3.0

Instance details

Methods

def :: EMDOpts v n a