Numeric.EMD.Internal.Spline

Description

Internal splining functionality exported for testing purposes only. This will likely go away in future versions, so please do not depend on this!

Documentation

data Spline a

1D Cubic spline

data SplineEnd a

End condition for spline

Constructors

SENotAKnot
"Not-a-knot" condition: third derivatives are continuous at endpoints. Default for matlab spline.

SENatural
"Natural" condition: curve becomes a straight line at endpoints.

SEClamped a a
"Clamped" condition: Slope of curves at endpoints are explicitly given.

Since: 0.1.2.0

makeSpline

Arguments

:: (Ord a, Fractional a)

=> SplineEnd a

-> Map a a
(x, y)

-> Maybe (Spline a)

Build a cubic spline based on control points using given end conditions (not-a-knot, or natural)

https://en.wikipedia.org/wiki/Spline_interpolation

sampleSpline :: (Fractional a, Ord a) => Spline a -> a -> a

Sample a spline at a given point.