Numeric.EMD.Unsized

Description

Interface of Numeric.EMD re-exported in a non-typesafe "unsized" form. Can be more convenient in certain situations, but Numeric.EMD is recommended and preferred.

Documentation

emd :: (Vector v a, Fractional a, Ord a) => EMDOpts a -> v a -> Maybe (EMD v a)

EMD decomposition (Hilbert-Huang Transform) of a given time series with a given sifting stop condition.

Returns Nothing if given an empty vector.

See emd for a type-safe version with guaruntees on the output vector sizes.

emdTrace :: (Vector v a, Fractional a, Ord a, MonadIO m) => EMDOpts a -> v a -> m (Maybe (EMD v a))

emd, but tracing results to stdout as IMFs are found. Useful for debugging to see how long each step is taking.

Returns Nothing if given an empty vector.

emd' :: (Vector v a, Ord a, Fractional a, Applicative m) => (SiftResult v a -> m r) -> EMDOpts a -> v a -> m (Maybe (EMD v a))

emd with a callback for each found IMF.

Returns Nothing if given an empty vector.

data EMD v a

An EMD v a is a Hilbert-Huang transform of a time series with items of type a stored in a vector v.

Constructors

EMD

Fields

emdIMFs :: ![v a]

emdResidual :: !(v a)

data EMDOpts a

Options for EMD composition.

Constructors

EO

Fields

eoSiftCondition :: SiftCondition a

stop condition for sifting

eoSplineEnd :: SplineEnd a

end conditions for envelope splines

eoBoundaryHandler :: Maybe BoundaryHandler

process for handling boundary

defaultEO :: Fractional a => EMDOpts a

Default EMDOpts

data BoundaryHandler

Constructors

BHClamp
Clamp envelope at end points (Matlab implementation)

BHSymmetric
Extend boundaries symmetrically

data SiftCondition a

Stop conditions for sifting process

Constructors

SCStdDev !a
Stop using standard SD method

SCTimes !Int
Stop after a fixed number of sifting iterations

SCOr (SiftCondition a) (SiftCondition a)
One or the other

SCAnd (SiftCondition a) (SiftCondition a)
Stop when both conditions are met

defaultSC :: Fractional a => SiftCondition a

Default SiftCondition

data SplineEnd a

End condition for spline

Constructors

SENotAKnot

SENatural

SEClamped a a

sift :: (Vector v a, Fractional a, Ord a) => EMDOpts a -> v a -> Maybe (SiftResult v a)

Iterated sifting process, used to produce either an IMF or a residual.

Returns Nothing if given an empty vector.

data SiftResult v a

The result of a sifting operation. Each sift either yields a residual, or a new IMF.

Constructors

SRResidual !(v a)

SRIMF !(v a) !Int
number of iterations