Numeric.EMD

Documentation

emd :: (Vector v a, KnownNat n, Fractional a, Ord a) => EMDOpts a -> Vector v (n + 2) a -> EMD v (n + 2) a

EMD decomposition (Hilbert-Huang Transform) of a given time series with a given sifting stop condition.

emdTrace :: (Vector v a, KnownNat n, Fractional a, Ord a, MonadIO m) => EMDOpts a -> Vector v (n + 2) a -> m (EMD v (n + 2) a)

emd, but tracing results to stdout as IMFs are found. Useful for debugging to see how long each step is taking.

emd' :: (Vector v a, KnownNat n, Fractional a, Ord a, Applicative m) => (SiftResult v (n + 2) a -> m r) -> EMDOpts a -> Vector v (n + 2) a -> m (EMD v (n + 2) a)

emd with a callback for each found IMF.

data EMD v n a

An EMD v n a is a Hilbert-Huang transform of a time series with n items of type a stored in a vector v.

Constructors

EMD

Fields

emdIMFs :: ![Vector v n a]

emdResidual :: !(Vector v n a)

data EMDOpts a

Options for EMD composition.

Constructors

EO

Fields

eoSiftCondition :: SiftCondition a

stop condition for sifting

eoSplineEnd :: SplineEnd

end conditions for envelope splines

eoClampEnvelope :: Bool

if True, use time series endpoints as part of min and max envelopes

defaultEO :: Fractional a => EMDOpts a

Default EMDOpts

data SiftCondition a

Stop conditions for sifting process

Constructors

SCStdDev a
Stop using standard SD method

SCTimes Int
Stop after a fixed number of iterations

SCOr (SiftCondition a) (SiftCondition a)
one or the other

SCAnd (SiftCondition a) (SiftCondition a)
both conditions met

defaultSC :: Fractional a => SiftCondition a

Default SiftCondition

sift :: (Vector v a, KnownNat n, Fractional a, Ord a) => EMDOpts a -> Vector v (n + 2) a -> SiftResult v (n + 2) a

Iterated sifting process.

Debug

envelopes :: (Vector v a, KnownNat n, Fractional a, Ord a) => SplineEnd -> Bool -> Vector v (n + 2) a -> Maybe (Vector v (n + 2) a, Vector v (n + 2) a)

Returns cubic splines of local minimums and maximums. Returns Nothing if there are not enough local minimum or maximums to create the splines.