Numeric.EMD

Description

Empirical Mode Decomposition in pure Haskell.

Main interface is emd, with defaultEO. A tracing version that outputs a log to stdout is also available, as emdTrace. This can be used to help track down a specific IMF that might be taking more time than desired.

This package uses "sized vectors" as its main interface, to ensure:

The resulting EMD contains IMFs that are all the same length as the input vector

We provide a vector of size of at least one.

There are many functions to convert unsized vectors to sized vectors in Data.Vector.Sized and associated modules, including toSized (for when you know the size at compile-time) and withSized (for when you don't).

Empirical Mode Decomposition

emd :: (Vector v a, KnownNat n, Fractional a, Ord a) => EMDOpts a -> Vector v (n + 1) a -> EMD v (n + 1) a

EMD decomposition of a given time series with a given sifting stop condition.

Takes a sized vector to ensure that:

The resulting EMD contains IMFs that are all the same length as the input vector

We provide a vector of size of at least one.

emdTrace :: (Vector v a, KnownNat n, Fractional a, Ord a, MonadIO m) => EMDOpts a -> Vector v (n + 1) a -> m (EMD v (n + 1) a)

emd, but tracing results to stdout as IMFs are found. Useful for debugging to see how long each step is taking.

emd' :: (Vector v a, KnownNat n, Fractional a, Ord a, Applicative m) => (SiftResult v (n + 1) a -> m r) -> EMDOpts a -> Vector v (n + 1) a -> m (EMD v (n + 1) a)

emd with a callback for each found IMF.

iemd :: (Vector v a, Num a) => EMD v n a -> Vector v n a

Collapse an EMD back into its original time series. Should be a left-inverse to emd: using iemd on the result of emd should give back the original vector.

Since: 0.1.5.0

data EMD v n a

An EMD v n a is an Empirical Mode Decomposition of a time series with n items of type a stored in a vector v.

The component-wise sum of emdIMFs and emdResidual should yield exactly the original series (see iemd).

Constructors

EMD

Fields

emdIMFs :: ![Vector v n a]

emdResidual :: !(Vector v n a)

data EMDOpts a

Options for EMD composition.

Constructors

EO

Fields

eoSiftCondition :: SiftCondition a

stop condition for sifting

eoSplineEnd :: SplineEnd a

end conditions for envelope splines

eoBoundaryHandler :: Maybe BoundaryHandler

process for handling boundary

defaultEO :: Fractional a => EMDOpts a

Default EMDOpts

data BoundaryHandler

Constructors

BHClamp
Clamp envelope at end points (Matlab implementation)

BHSymmetric
Extend boundaries symmetrically

data SiftCondition a

Stop conditions for sifting process

Data type is lazy in its fields, so this infinite data type:

nTimes n = SCTimes n SCOr nTimes (n + 1)

will be treated identically as:

nTimes = SCTimes

Constructors

SCStdDev !a
Stop using standard SD method

SCTimes !Int
Stop after a fixed number of sifting iterations

SCOr (SiftCondition a) (SiftCondition a)
One or the other

SCAnd (SiftCondition a) (SiftCondition a)
Stop when both conditions are met

defaultSC :: Fractional a => SiftCondition a

Default SiftCondition

data SplineEnd a

End condition for spline

Constructors

SENotAKnot
"Not-a-knot" condition: third derivatives are continuous at endpoints. Default for matlab spline.

SENatural
"Natural" condition: curve becomes a straight line at endpoints.

SEClamped a a
"Clamped" condition: Slope of curves at endpoints are explicitly given.

Since: 0.1.2.0

Internal

sift :: forall v n a. (Vector v a, KnownNat n, Fractional a, Ord a) => EMDOpts a -> Vector v (n + 1) a -> SiftResult v (n + 1) a

Iterated sifting process, used to produce either an IMF or a residual.

data SiftResult v n a

The result of a sifting operation. Each sift either yields a residual, or a new IMF.

Constructors

SRResidual !(Vector v n a)

SRIMF !(Vector v n a) !Int
number of sifting iterations

envelopes :: (Vector v a, KnownNat n, Fractional a, Ord a) => SplineEnd a -> Maybe BoundaryHandler -> Vector v (n + 1) a -> Maybe (Vector v (n + 1) a, Vector v (n + 1) a)

Returns cubic splines of local minimums and maximums. Returns Nothing if there are not enough local minimum or maximums to create the splines.