Numeric.HHT

Documentation

hhtEmd :: forall v n a. (Vector v a, KnownNat n, RealFloat a) => EMD v (n + 1) a -> HHT v n a

hht :: forall v n a. (Vector v a, KnownNat n, RealFloat a) => EMDOpts a -> Vector v (n + 1) a -> HHT v n a

marginal :: forall v n a. (Vector v a, KnownNat n, Ord a, Num a) => HHT v n a -> Map a a

instantaneousEnergy :: forall v n a. (Vector v a, KnownNat n, Num a) => HHT v n a -> Vector v n a

newtype HHT v n a

Constructors

HHT

Fields

hhtLines :: [HHTLine v n a]

data HHTLine v n a

Constructors

HHTLine

Fields

hlMags :: !(Vector v n a)

hlFreqs :: !(Vector v n a)

data EMDOpts a

Options for EMD composition.

Constructors

EO

Fields

eoSiftCondition :: SiftCondition a

stop condition for sifting

eoSplineEnd :: SplineEnd a

end conditions for envelope splines

eoClampEnvelope :: Bool

if True, use time series endpoints as part of min and max envelopes

defaultEO :: Fractional a => EMDOpts a

Default EMDOpts

data SiftCondition a

Stop conditions for sifting process

Constructors

SCStdDev !a
Stop using standard SD method

SCTimes !Int
Stop after a fixed number of sifting iterations

SCOr (SiftCondition a) (SiftCondition a)
one or the other

SCAnd (SiftCondition a) (SiftCondition a)
both conditions met

defaultSC :: Fractional a => SiftCondition a

Default SiftCondition

data SplineEnd a

End condition for spline

Constructors

SENotAKnot

SENatural

SEClamped a a

Hilbert transforms (internal usage)

hilbert :: forall v n a. (Vector v a, Vector v (Complex a), KnownNat n, Floating a) => Vector v n a -> Vector v n (Complex a)

Real part is original series and imaginary part is hilbert transformed series.

hilbert' :: forall v n a. (Vector v a, KnownNat n, Floating a) => Vector v n a -> Vector v n a

Can be made faster using an FFT and iFFT combo