Numeric.HHT

Description

Hilbert-Huang transform in pure Haskell.

The main data type is HHT, which can be generated using hht or hhtEmd. See Numeric.EMD for information on why this module uses "sized vectors", and how to convert unsized vectors to sized vectors.

Since: 0.1.2.0

Hilbert-Huang Transform

data HHT v n a

A Hilbert-Huang Transform. An HHT v n a is a Hilbert-Huang transform of an n-item time series of items of type a represented using vector v.

Create using hht or hhtEmd.

Constructors

HHT

Fields

hhtLines :: [HHTLine v n a]

Skeleton lines corresponding to each IMF

hhtResidual :: Vector v (n + 1) a

Residual from EMD

Since: 0.1.9.0

data HHTLine v n a

A Hilbert Trasnform of a given IMF, given as a "skeleton line".

Constructors

HHTLine

Fields

hlMags :: !(Vector v (n + 1) a)

IMF HHT Magnitude as a time series.

It may be useful to "zip" this vector with hlFreqs. To do this, use a function like init or tail to make these two vectors contain the same length, or weaken/shift to make indices in hlFreqs usable as indices in hlMags.

Prior to v0.1.9.0, this was a length-n vector, just like hlFreqs. To get the same behavior, use init on this new field's value.

hlFreqs :: !(Vector v n a)

IMF HHT instantaneous frequency as a time series (between 0 and 1).

In reality, these frequencies are the frequencies "in between" each step in hlMags.

hlInitPhase :: !a

Initial phase of skeleton line (between -pi and pi)

Since: 0.1.9.0

hhtEmd :: forall v n a. (Vector v a, Vector v (Complex a), KnownNat n, FFTWReal a) => EMD v (n + 1) a -> HHT v n a

Compute the Hilbert-Huang transform from a given Empirical Mode Decomposition.

hht :: forall v n a. (Vector v a, Vector v (Complex a), KnownNat n, FFTWReal a) => EMDOpts a -> Vector v (n + 1) a -> HHT v n a

Directly compute the Hilbert-Huang transform of a given time series. Essentially is a composition of hhtEmd and emd. See hhtEmd for a more flexible version.

ihhtEmd :: (Vector v a, Floating a) => HHT v n a -> EMD v (n + 1) a

Invert a Hilbert-Huang transform back to an Empirical Mode Decomposition

Since: 0.1.9.0

ihht :: (Vector v a, Floating a) => HHT v n a -> Vector v (n + 1) a

Construct a time series correpsonding to its hilbert-huang transform.

Since: 0.1.9.0

Hilbert-Huang Spectrum

hhtSpectrum

Arguments

:: forall v n a k. (Vector v a, KnownNat n, Ord k, Num a)

=> (a -> k)
binning function. takes rev/tick freq between 0 and 1.

-> HHT v n a

-> Vector n (Map k a)

Compute the full Hilbert-Huang Transform spectrum. At each timestep is a sparse map of frequency components and their respective magnitudes. Frequencies not in the map are considered to be zero.

Takes a "binning" function to allow you to specify how specific you want your frequencies to be.

See hhtSparseSpetrum for a sparser version, and hhtDenseSpectrum for a denser version.

hhtSparseSpectrum

Arguments

:: forall v n a k. (Vector v a, KnownNat n, Ord k, Num a)

=> (a -> k)
binning function. takes rev/tick freq between 0 and 1.

-> HHT v n a

-> Map (Finite n, k) a

A sparser vesion of hhtSpectrum. Compute the full Hilbert-Huang Transform spectrum. Returns a sparse matrix representing the power at each time step (the Finite n) and frequency (the k).

Takes a "binning" function to allow you to specify how specific you want your frequencies to be.

Since: 0.1.4.0

hhtDenseSpectrum

Arguments

:: forall v n m a. (Vector v a, KnownNat n, KnownNat m, Num a)

=> (a -> Finite m)
binning function. takes rev/tick freq between 0 and 1.

-> HHT v n a

-> Vector n (Vector m a)

A denser version of hhtSpectrum. Compute the full Hilbert-Huang Transform spectrum, returning a dense matrix (as a vector of vectors) representing the power at each time step and each frequency.

Takes a "binning" function that maps a frequency to one of m discrete slots, for accumulation in the dense matrix.

Since: 0.1.4.0

Properties of spectrum

meanMarginal

Arguments

:: forall v n a k. (Vector v a, KnownNat n, Ord k, Fractional a)

=> (a -> k)
binning function. takes rev/tick freq between 0 and 1.

-> HHT v n a

-> Map k a

Compute the mean marginal spectrum given a Hilbert-Huang Transform. It is similar to a Fourier Transform; it provides the "total power" over the entire time series for each frequency component, averaged over the length of the time series.

A binning function is accepted to allow you to specify how specific you want your frequencies to be.

Since: 0.1.8.0

marginal

Arguments

:: forall v n a k. (Vector v a, KnownNat n, Ord k, Num a)

=> (a -> k)
binning function. takes rev/tick freq between 0 and 1.

-> HHT v n a

-> Map k a

Compute the marginal spectrum given a Hilbert-Huang Transform. It provides the "total power" over the entire time series for each frequency component. See meanMarginal for a version that averages over the length of the time series, making it more close in nature to the purpose of a Fourier Transform.

A binning function is accepted to allow you to specify how specific you want your frequencies to be.

instantaneousEnergy :: forall v n a. (Vector v a, KnownNat n, Num a) => HHT v n a -> Vector v n a

Compute the instantaneous energy of the time series at every step via the Hilbert-Huang Transform.

degreeOfStationarity

Arguments

:: forall v n a k. (Vector v a, KnownNat n, Ord k, Fractional a, Eq a)

=> (a -> k)
binning function. takes rev/tick freq between 0 and 1.

-> HHT v n a

-> Map k a

Degree of stationarity, as a function of frequency.

expectedFreq :: forall v n a. (Vector v a, KnownNat n, Fractional a) => HHT v n a -> Vector v n a

Returns the "expected value" of frequency at each time step, calculated as a weighted average of all contributions at every frequency at that time step.

Since: 0.1.4.0

dominantFreq :: forall v n a. (Vector v a, KnownNat n, Ord a) => HHT v n a -> Vector v n a

Returns the dominant frequency (frequency with largest magnitude contribution) at each time step.

Since: 0.1.4.0

foldFreq

Arguments

:: forall v u n a b c. (Vector v a, Vector u c, KnownNat n, Monoid b)

=> (a -> a -> b)
Combining function, taking frequency, then magnitude

-> (b -> c)
Projecting function

-> HHT v n a

-> Vector u n c

Fold and collapse a Hilbert-Huang transform along the frequency axis at each step in time along some monoid.

Since: 0.1.8.0

Options

data EMDOpts a

Options for EMD composition.

Constructors

EO

Fields

eoSiftCondition :: SiftCondition a

stop condition for sifting

eoSplineEnd :: SplineEnd a

end conditions for envelope splines

eoBoundaryHandler :: Maybe BoundaryHandler

process for handling boundary

defaultEO :: Fractional a => EMDOpts a

Default EMDOpts

data BoundaryHandler

Constructors

BHClamp
Clamp envelope at end points (Matlab implementation)

BHSymmetric
Extend boundaries symmetrically

data SiftCondition a

Stop conditions for sifting process

Data type is lazy in its fields, so this infinite data type:

nTimes n = SCTimes n SCOr nTimes (n + 1)

will be treated identically as:

nTimes = SCTimes

Constructors

SCStdDev !a
Stop using standard SD method

SCTimes !Int
Stop after a fixed number of sifting iterations

SCOr (SiftCondition a) (SiftCondition a)
One or the other

SCAnd (SiftCondition a) (SiftCondition a)
Stop when both conditions are met

defaultSC :: Fractional a => SiftCondition a

Default SiftCondition

data SplineEnd a

End condition for spline

Constructors

SENotAKnot
"Not-a-knot" condition: third derivatives are continuous at endpoints. Default for matlab spline.

SENatural
"Natural" condition: curve becomes a straight line at endpoints.

SEClamped a a
"Clamped" condition: Slope of curves at endpoints are explicitly given.

Since: 0.1.2.0

Hilbert transforms (internal usage)

hilbert :: forall v n a. (Vector v a, Vector v (Complex a), KnownNat n, FFTWReal a) => Vector v n a -> Vector v n (Complex a)

Real part is original series and imaginary part is hilbert transformed series. Creates a "helical" form of the original series that rotates along the complex plane.

Note that since 0.1.7.0, this uses the same algorithm as the matlab implementation https://www.mathworks.com/help/signal/ref/hilbert.html

hilbertIm :: forall v n a. (Vector v a, Vector v (Complex a), KnownNat n, FFTWReal a) => Vector v n a -> Vector v n a

Hilbert transformed series. Essentially the same series, but phase-shifted 90 degrees. Is so-named because it is the "imaginary part" of the proper hilbert transform, hilbert.

Note that since 0.1.7.0, this uses the same algorithm as the matlab implementation https://www.mathworks.com/help/signal/ref/hilbert.html

hilbertPolar :: forall v n a. (Vector v a, Vector v (Complex a), KnownNat n, FFTWReal a) => Vector v (n + 1) a -> (Vector v (n + 1) a, Vector v (n + 1) a)

The polar form of hilbert: returns the magnitude and phase of the discrete hilbert transform of a series.

The computation of magnitude is unique, but computing phase gives us some ambiguity. The interpretation of the hilbert transform for instantaneous frequency is that the original series "spirals" around the complex plane as time progresses, like a helix. So, we impose a constraint on the phase to uniquely determine it: \(\phi_{t+1}\) is the minimal valid phase such that \(\phi_{t+1} \geq \phi_{t}\). This enforces the phase to be monotonically increasing at the slowest possible detectable rate.

Since: 0.1.6.0

hilbertMagFreq :: forall v n a. (Vector v a, Vector v (Complex a), KnownNat n, FFTWReal a) => Vector v (n + 1) a -> (Vector v (n + 1) a, (Vector v n a, a))

Given a time series, return a time series of the magnitude of the hilbert transform and the frequency of the hilbert transform, in units of revolutions per tick. Is only expected to taken in proper/legal IMFs.

The frequency will always be between 0 and 1, since we can't determine anything faster given the discretization, and we exclude negative values as physically unmeaningful for an IMF.